fixup! refactor(dev-infra): migrate github api in GitClient to rely on GithubClient
